Minor refactoring, moving response_handler impl into the spi.cpp.
